KDW006 masterlist fix INPUT_UNIT

+++ nts.uk/uk.at/at.file/nts.uk.file.at.app/src/main/java/nts/uk/file/at/app/export/attendanceitemprepare/ControlOfAttendanceItemsDtoExcel.java
@@ -20,6 +20,6 @@ public class ControlOfAttendanceItemsDtoExcel {
 	private String headerBgColorOfDailyPer;
 
 	/**時間項目の入力単位*/
-	private Integer inputUnitOfTimeItem;
+	private Float inputUnitOfTimeItem;
 	
 }
