+++ nts.uk/uk.at/at.file/nts.uk.file.at.infra/src/main/java/nts/uk/file/at/infra/export/attendanceitemprepare/JpaControlOfAttItemsRepo.java
@@ -32,13 +32,13 @@ public class JpaControlOfAttItemsRepo extends JpaRepository implements ControlOf
 
 	private void putRowToResult(Map<Integer, ControlOfAttendanceItemsDtoExcel> listControl, Object[] x) {
 		Integer id = 0;
-		Integer unitItem = 0;
+		Float unitItem = 0f;
 		String headerBackground = (String) x[2];
 		if(x[0] !=null){
 			id = ((BigDecimal) x[0]).intValue();
 		}
 		if(x[1] !=null){
-			unitItem =((BigDecimal) x[1]).intValue();
+			unitItem =((BigDecimal) x[1]).floatValue();
 		}
 		ControlOfAttendanceItemsDtoExcel control = 
 				new ControlOfAttendanceItemsDtoExcel(id, headerBackground,unitItem );
